Fix Nix implicit includes for clang-tidy/IWYU, add warning flags and GL debug context
CMake strips Nix-injected implicit include paths from compile commands,
which breaks clang-tidy and IWYU (they don't use the Nix clang wrapper).
Fix by re-adding CMAKE_CXX_IMPLICIT_INCLUDE_DIRECTORIES as extra args.

Also adds Phase 1 items from the design doc:
- Extra compiler warnings (-Wshadow, -Wformat=2, -Wimplicit-fallthrough)
  for engine code only (vendored code exempt)
- GL debug context flag in debug builds
- Fix all new warnings and clang-tidy findings

+++ src/lua.cc
@@ -975,8 +975,8 @@ void Lua::Update(float t, float dt) {
   lua_getglobal(state_, "_Game");
   lua_getfield(state_, -1, "update");
   lua_insert(state_, -2);
-  lua_pushnumber(state_, t);
-  lua_pushnumber(state_, dt);
+  lua_pushnumber(state_, static_cast<double>(t));
+  lua_pushnumber(state_, static_cast<double>(dt));
   if (lua_pcall(state_, 3, 0, traceback_handler_)) {
     lua_error(state_);
     return;
@@ -1111,10 +1111,10 @@ void Lua::HandleMouseMoved(FVec2 pos, FVec2 delta) {
     return;
   }
   lua_insert(state_, -2);
-  lua_pushnumber(state_, pos.x);
-  lua_pushnumber(state_, pos.y);
-  lua_pushnumber(state_, delta.x);
-  lua_pushnumber(state_, delta.y);
+  lua_pushnumber(state_, static_cast<double>(pos.x));
+  lua_pushnumber(state_, static_cast<double>(pos.y));
+  lua_pushnumber(state_, static_cast<double>(delta.x));
+  lua_pushnumber(state_, static_cast<double>(delta.y));
   if (lua_pcall(state_, 5, 0, traceback_handler_)) {
     lua_error(state_);
     return;
